Add only_admins and some extra functionality on user join/leave

+++ tg/handlers/chapa/delete.py
@@ -1,9 +1,7 @@
-from typing import List
-
-from telegram import Update, Bot, Message
+from telegram import Update, Bot, Message, Chat
 
 from models.schedule import ScheduledTask
-from tg.utils.decorators import log_update
+from tg.decorators import log_update, only_admins
 from models.phrase import Phrase
 
 
@@ -16,16 +14,43 @@ def usage(update: Update) -> Message:
     )
 
 
-def delete_all_chapas(tasks: List[ScheduledTask], update: Update):
+@only_admins
+@log_update
+def handle_delete_chapas(bot: Bot, update: Update):
+    chat: Chat = update.effective_chat
+
+    tasks = ScheduledTask.get_tasks(chat_id=chat.id)
+    if len(tasks) == 0:
+        return update.effective_message.reply_text(
+            f"¡Pero si no te estoy dando la chapa, {Phrase.get_random_phrase()}!"
+        )
+
     for task in tasks:
         task.delete()
+
     update.effective_message.reply_text(
         f"Ya no te daré más chapas, ({len(tasks)} borradas) {Phrase.get_random_phrase()}.",
         quote=True,
     )
 
 
-def delete_one_chapa(tasks: List[ScheduledTask], chapa_id: int, update: Update):
+
+@only_admins
+@log_update
+def handle_delete_chapa(bot: Bot, update: Update):
+    tokens = update.effective_message.text.split(" ")
+    try:
+        cmd = tokens[0]
+        chapa_id = int(tokens[1]) - 1
+    except (IndexError, ValueError) as e:
+        return usage(update)
+
+    tasks = ScheduledTask.get_tasks(chat_id=update.effective_chat.id)
+    if len(tasks) == 0:
+        return update.effective_message.reply_text(
+            f"¡Pero si no te estoy dando la chapa, {Phrase.get_random_phrase()}!"
+        )
+
     try:
         task = tasks[chapa_id]
         task.delete()
@@ -33,27 +58,5 @@ def delete_one_chapa(tasks: List[ScheduledTask], chapa_id: int, update: Update):
         update.effective_message.reply_text(f"Te has pasado con el número, {Phrase.get_random_phrase()}.", quote=True)
     else:
         update.effective_message.reply_text(f"Ya no te daré esa chapa, {Phrase.get_random_phrase()}.", quote=True)
-
-
-@log_update
-def handle_delete_chapa(bot: Bot, update: Update):
-    tokens = update.effective_message.text.split(" ")
-    try:
-        breakpoint()
-        cmd = tokens[0]
-        if 'borrarchapas' in cmd:
-            chapa_id = None
-        else:
-            chapa_id = int(tokens[1]) - 1
-    except (IndexError, ValueError) as e:
-        return usage(update)
-
-    tasks = ScheduledTask.get_tasks(chat_id=update.effective_chat.id)
-    if len(tasks) == 0:
-        return update.effective_message.reply_text(f"¡Pero si no te estoy dando la chapa, {Phrase.get_random_phrase()}!")
-    if chapa_id is None:
-        delete_all_chapas(tasks, update)
-    else:
-        delete_one_chapa(tasks, chapa_id, update)
